Attempted other refactoring, given up
It is not nice that we do not have an optimal way
to distinguish leaf-like "Site" levels from the others
(That would require a composite pattern, maybe?)
This may be related to the implementation of
get_indices_tree_with_ghosts.
Leaving things as they are at the moment
(when using get_indices_tree_with_ghosts
on a tree with leaf, use truncate_tree afterwards).