API-V2: unified index lookup, fixed vector print
index_pullback and index_pushforward
now use the same underlying function,
that can be told to interpret the input as indices
and retrieve the key at certain index in the level (pullback)
or to interpret the input as keys,
and then retrieve the index at which
the key matches the right portion of the input.

This should remove the need
for renumbering the nodes at certain levels,
allowing condensing multiple transformation into one
(e.g., for eo_naive, which needs a flatten before).