API-V2: still not working
It seems that for this approach to work
many function needs to be reworked completely,
as there is an inconsistency between the meaning of indices
if transformations are applied to the tree
without renumbering.
Now the pushforward of the first transformation
gives indices for its competence range
(by design, this is what is required)
but the second transformation expects keys
for those levels that are not in its competence.
This is a fatal inconsistency.

It could still make sense to define the competence range
to be 0<=l<X, instead of Y<l<X.
This would still require renumbering,
but not beyond level X,
which should remove the need to have "global" memoisation.