removed GHOST, bodge for partition_children_...
Changes to partition_children_into_subtrees.
The problem with the "GHOST" node type is that
having multiple node types at the same level
messes up with the level swapping algorithm.
Having simply an empty tree messes up with the partitioning
(e.g., the HBB algorithm).
So copying in a pointer to an existing subtree,
strapping a {no_key} key on it, to prevent it from being used
(except in partitioning when the subsequent levels need it)

NOTE: this only affects open boundary conditions,
where there might be missing sites for the halos.